And another iteration
Enable 'isTrigger' field on TowerPrefab collider, hopefully OnMouseDown
will now trigger.
Re-enabled some RPC functions in Player.cs. They can only be resolved at
runtime, so you can remove them and still succesfully compile.
Towers and projectiles are now spawned on the server and their logic
(OnUpdate) is also handled there.
Ran the client in editor mode and discovered some errors trying to
initialize A* pathfinding stuff on the client. Since this only exists on
the server, it won't work. One warning still remains.
Made a start with the code to sell towers. Made a link between a Tile
and Tower object. Only thing left to do is to get a reference to the
GameField from the OnMouseDown in Tower. We need to think about some
global helper function or a logical place to put all our RPC calls in.